backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Vector database and embeddings
faiss-cpu>=1.8.0
sentence-transformers==2.3.1
torch>=2.2.0
numpy==1.26.3

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10

# NLP and text processing
langchain==0.1.4
langchain-community==0.0.16
openai==1.10.0
tiktoken==0.5.2

# HTTP client for FHIR
httpx==0.26.0
requests==2.31.0

# Background tasks
celery==5.3.6
redis==5.0.1

# Utilities
pydantic==2.7.4
pydantic-settings==2.1.0
